% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_scale_identifiability.R
\name{check_scale_identifiability}
\alias{check_scale_identifiability}
\title{Check if scaling (b) is identifiable yet, with a more sensitive rule}
\usage{
check_scale_identifiability(
  currentD,
  g_ref_fun,
  hyper,
  min_week = 20,
  g_range_thresh = 0.25,
  p_range_thresh = 0.05
)
}
\arguments{
\item{currentD}{data frame with newWeek, y, neg.}

\item{g_ref_fun}{reference spline on link scale.}

\item{hyper}{list from learn_alignment_hyperparams().}

\item{min_week}{do not allow scaling before this epi week in newWeek space.}

\item{g_range_thresh}{minimum range of g(u) (on link scale) to trust scaling.}

\item{p_range_thresh}{minimum range of crude positivity to trust scaling.}
}
\value{
list with allow_scale_rec (TRUE/FALSE) and diagnostics.
}
\description{
Check if scaling (b) is identifiable yet, with a more sensitive rule
}
