% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_multi_template.R
\name{.weighted_quantile}
\alias{.weighted_quantile}
\title{Weighted quantile (internal helper)}
\usage{
.weighted_quantile(x, w, prob)
}
\description{
Weighted quantile (internal helper)
}
\keyword{internal}
