% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_newWeek_from_week}
\alias{get_newWeek_from_week}
\title{Map surveillance week to newWeek index in a season}
\usage{
get_newWeek_from_week(season_df, week_vec)
}
\arguments{
\item{season_df}{data frame with columns `week` and `newWeek`}

\item{week_vec}{vector of surveillance weeks to map}
}
\value{
integer vector of newWeek indices
}
\description{
Map surveillance week to newWeek index in a season
}
