% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_refCurve.R
\name{load_refCurve}
\alias{load_refCurve}
\title{Load all flualign example objects}
\usage{
load_refCurve(envir = parent.frame())
}
\arguments{
\item{envir}{environment to load into. Defaults to parent.frame().}
}
\description{
Loads all objects stored in the bundled ref_curve.RData file
into the specified environment (default: caller).
}
