% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_g_ref_mu_se}
\alias{make_g_ref_mu_se}
\title{Build reference mean/SE function from GAM (link scale)}
\usage{
make_g_ref_mu_se(gam_obj)
}
\arguments{
\item{gam_obj}{a fitted mgcv::gam (or gamm4::$gam)}
}
\value{
a function f(u) that returns list(mu = ..., se = ...) on link scale
}
\description{
Build reference mean/SE function from GAM (link scale)
}
