% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m1_reference_helpers.R
\name{set_reference}
\alias{set_reference}
\title{Set reference closures globally for convenience}
\usage{
set_reference(gam_obj, grid)
}
\description{
Set reference closures globally for convenience
}
