% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospective_training.R
\name{stage2_ramp_weight}
\alias{stage2_ramp_weight}
\title{Stage-2 ramp weights (internal)}
\usage{
stage2_ramp_weight(t_since, K = 3L)
}
\description{
Stage-2 ramp weights (internal)
}
\keyword{internal}
